user-service/requirements.txt
# Copyright 2024 Taufik Hidayat authors.
# 
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# 
#      http://www.apache.org/licenses/LICENSE-2.0
# 
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Core
cel-python==0.3.0
grpcio==1.67.1
grpcio-tools==1.67.1
googleapis-common-protos==1.65.0
grpcio-status==1.67.1
bcrypt==4.2.0
PyJWT==2.9.0
asyncpg==0.30.0
pydantic==2.13.1
pydantic[email]==2.13.1

# Dev
isort==5.13.2
ipython==8.29.0
mypy===1.3.0
ruff==0.5.3
pytest==9.0.3
pytest-asyncio==0.22.0
psycopg2-binary==2.9.10
yoyo-migrations==9.0.0
email-validator>=2.0.0,<3.0.0

# Observability (LGTM)
python-json-logger==2.0.7
prometheus_client==0.16.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-grpc==0.49b0
